pyproject.toml
[tool.poetry]
name = "threat-db"
version = "0.1.0"
description = "A graphql server for vulnerabilities powered by dgraph"
authors = ["Team ngcloudsec <cloud@ngcloud.io>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "threat_db"}]
homepage = ""
repository = ""
keywords = []
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Utilities",
    "Topic :: Security",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]

[tool.poetry.scripts]
threat_db = 'threat_db.cli:main'

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
pydgraph = "^21.3.2"
grpcio = "^1.51.1"
rich = "^12.6.0"
protobuf = "3.20.1"
packageurl-python = "^0.10.4"
gql = {extras = ["all"], version = "^3.4.0"}
flask = "^2.2.2"
uwsgi = "^2.0.21"
orjson = "^3.8.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
pyinstaller = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
